Refactor : WinningResult Dto 생성
