+++ src/main/java/lotto/model/WinningResult.java
@@ -4,22 +4,28 @@ import java.util.Map;
 import java.util.function.ToDoubleFunction;
 
 public class WinningResult {
-    private final Map<Rank, Integer> winningResult;
 
-    WinningResult(final Map<Rank, Integer> winningResult) {
+    private final Map<Rank, Integer> winningResult;
+    private final double profit;
+
+    public WinningResult(final Map<Rank, Integer> winningResult, final LottoMoney lottoMoney) {
         this.winningResult = winningResult;
+        this.profit = calculateProfit(lottoMoney);
     }
 
     public int getRankCount(final Rank rank) {
         return winningResult.get(rank);
     }
 
-    public double calculateProfit(final LottoMoney gain) {
+    public double getProfit() {
+        return this.profit;
+    }
+
+    private double calculateProfit(final LottoMoney lottoMoney) {
         double sum = winningResult.keySet().stream()
-                .mapToDouble(getRankRewards())
-                .sum();
-        int lottoMoney = gain.getLottoMoney();
-        return sum / lottoMoney;
+            .mapToDouble(getRankRewards())
+            .sum();
+        return sum / lottoMoney.getLottoMoney();
     }
 
     private ToDoubleFunction<Rank> getRankRewards() {
